drawer and main list css
