+++ frontend/src/Components/CustomList/customList.css
@@ -1,5 +1,7 @@
 .custom-list-section {
-    margin: 64px 16px 0px;
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 64px 16px 0px;    
 }
 
 .list-header{
@@ -9,7 +11,23 @@
 
 #addNewButton {
     width: 100%;
-    background-color: antiquewhite;
+    padding: 8px;
+    color: brown;
+    font-size: 16px;
+    font-weight: bold;
+    background-color: rgb(233, 150, 122, 0.6);
+    border-radius: 20px;
+    border: 1px solid brown;
+}
+
+#addNewButton:hover {
+    cursor: pointer;
+    box-shadow: 4px 4px 8px grey;
+}
+
+#addNewButton:active {
+    cursor: pointer;
+    box-shadow: 0px 0px 20px grey inset;
 }
 
 .listItems {
@@ -18,12 +36,8 @@
     align-items: center;
     margin: 16px 0;
     padding: 0 16px;
-    background-color: antiquewhite;
+    background-color: wheat;
     border-radius: 20px;
-}
-
-.listItems:hover {
-    cursor: pointer;
 }
 
 .listItemHeader {
@@ -32,15 +46,28 @@
     letter-spacing: 1px;
 }
 
-.deleteIcon {
-    margin: 12px 0;
-    padding: 12px;
-    border-radius: 50%;
+.editIcon {
+    margin: 12px 8px 12px 0;
 }
 
-.deleteIcon:hover {
+.deleteIcon {
+    margin: 12px 0;
+}
+
+.deleteIcon, .editIcon {
+    padding: 8px;
+    border-radius: 12px;
+    background-color: #ffffff66;
+}
+
+.deleteIcon:hover, .editIcon:hover {
     color: white;
+    cursor: pointer;
     font-weight: bold;
     background-color: salmon;
     box-shadow: 0px 0px 8px firebrick;
 }
+
+.listIconGroup {
+    display: flex;
+  }
